Sarah updates current (#49)
* Updates to Enviroment

* Cliff and Trees update

* Influence + Combat UI update

Virtuoso Bar moved and shrunk to fit into upper right corner. Now can only influence allies and not the player character.

* Doors Added to Village

* update

* Influence disabled update

Influence is disabled when there is only one ally character in the party laive.

* Code Cleanup

* Added Townspeople

* Townspeople and window update

Added glow to some windows and added in more npcs. Also tried to diversify the npc models with different textures.

* Update EditorUserSettings.asset

* minor update

* village managers + dialogue

Added in manager scripts and some dialogue to the village.

* UI Now Can run in Village

* The Herbalist can now say things

* Basic Dialogue Model Added for Town

* DialogueModelUpdates

* Minor update to fix ui bug in tavern

* Small Fix

* Update default-2021.dwlt

---------

Co-authored-by: kdavidmitchell <kdavidmitchell@gmail.com>
Co-authored-by: kdavidmitchell <30902200+kdavidmitchell@users.noreply.github.com>